Use floats instead of doubles for all Faust UI calculations & drawing. Switch from custom `Point` class to `ImGui::ImVec2`

+++ src/UI/Faust/Device.h
@@ -9,15 +9,15 @@ enum { kLeftRight = 1, kRightLeft = -1 };
 class Device {
 public:
     virtual ~Device() = default;
-    virtual void rect(double x, double y, double l, double h, const string &color, const string &link) = 0;
-    virtual void triangle(double x, double y, double l, double h, const string &color, int orientation, const string &link) = 0;
-    virtual void circle(double x, double y, double radius) = 0;
-    virtual void square(double x, double y, double dim) = 0;
-    virtual void arrow(double x, double y, double rotation, int orientation) = 0;
-    virtual void line(double x1, double y1, double x2, double y2) = 0;
-    virtual void dasharray(double x1, double y1, double x2, double y2) = 0;
-    virtual void text(double x, double y, const char *name, const string &link) = 0;
-    virtual void label(double x, double y, const char *name) = 0;
-    virtual void dot(double x, double y, int orientation) = 0;
-    virtual void Error(const char *message, const char *reason, int nb_error, double x, double y, double width) = 0;
+    virtual void rect(float x, float y, float l, float h, const string &color, const string &link) = 0;
+    virtual void triangle(float x, float y, float l, float h, const string &color, int orientation, const string &link) = 0;
+    virtual void circle(float x, float y, float radius) = 0;
+    virtual void square(float x, float y, float dim) = 0;
+    virtual void arrow(float x, float y, float rotation, int orientation) = 0;
+    virtual void line(float x1, float y1, float x2, float y2) = 0;
+    virtual void dasharray(float x1, float y1, float x2, float y2) = 0;
+    virtual void text(float x, float y, const char *name, const string &link) = 0;
+    virtual void label(float x, float y, const char *name) = 0;
+    virtual void dot(float x, float y, int orientation) = 0;
+    virtual void Error(const char *message, const char *reason, int nb_error, float x, float y, float width) = 0;
 };
